sub-rpc, prototype support in zerorpc cli.
If the environment variable ZERORPC_CONNECT_TO is defined, zerorpc's cli
will use its content as the default endpoint to connect to.

When opening a server context from zerorpc's cli, a subshell is opened,
with ZERORPC_CONNECT_TO pointing to a temporary ipc.

The IPC is a proxy to the opened context. This proxy is really simple
and stupid right now, and doesn't support more than one client at a
time (it shouldn't be too much of a problem for most cases anyway).